Adds portions of enigma to code it module
