+++ lib/shiftgenerator.rb
@@ -43,10 +43,6 @@ module ShiftGenerator
     end
   end
 
-  # def character_set
-  #  ("a".."z").to_a << " "
-  # end
-
   def split_characters(words)
     split_message = words.downcase.split(//)
     split_message
